# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Friday Food Club API — drops, orders, chefs, admin, subscriptions
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: dropId
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a pre-order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed with inventory state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}/cancel:
    patch:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel a pending order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelResult"

  # ── Chefs ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, dropId, userId, price, status, orderedAt]
      properties:
        id:             { type: string }
        dropId:         { type: string }
        userId:         { type: string }
        price:          { type: number }
        effectivePrice: { type: number }
        isMemberOrder:  { type: boolean }
        status:         { type: string }
        orderedAt:      { type: string }
        paymentMethod:  { type: string, enum: [DIGITAL, CASH], description: "DIGITAL = pre-escrow; CASH = pay at pickup" }
        escrowStatus:   { type: ["string", "null"], enum: [HELD, RELEASED, CASH, CASH_RECONCILED, CANCELLED, null] }
        pickupToken:    { type: ["string", "null"] }
        dropTitle:      { type: ["string", "null"] }
        mealSlot:       { type: ["string", "null"] }
        chefName:       { type: ["string", "null"] }
        expiresAt:      { type: ["string", "null"] }
        inventory:      { type: ["integer", "null"] }
        currentOrders:  { type: ["integer", "null"] }
        dropStatus:     { type: ["string", "null"] }
        pickupLocation: { type: ["string", "null"] }
    OrderList:
      type: object
      required: [orders, total]
      properties:
        orders: { type: array, items: { $ref: "#/components/schemas/Order" } }
        total:  { type: integer }
    OrderInput:
      type: object
      required: [userId, dropId]
      properties:
        userId:        { type: string }
        dropId:        { type: string }
        paymentMethod: { type: string, enum: [DIGITAL, CASH], default: DIGITAL }
    OrderResult:
      type: object
      required: [order, drop]
      properties:
        order:
          type: object
          properties:
            id:             { type: string }
            dropId:         { type: string }
            userId:         { type: string }
            price:          { type: number }
            effectivePrice: { type: number }
            isMemberOrder:  { type: boolean }
            status:         { type: string }
            orderedAt:      { type: string }
            memberDiscount: { type: number }
            paymentMethod:  { type: string, enum: [DIGITAL, CASH] }
            escrowStatus:   { type: string, enum: [HELD, RELEASED, CASH, CASH_RECONCILED, CANCELLED] }
            pickupToken:    { type: ["string", "null"] }
        drop:
          type: object
          properties:
            id:            { type: string }
            currentOrders: { type: integer }
            inventory:     { type: integer }
            remaining:     { type: integer }
            status:        { type: string }
            justUnlocked:  { type: boolean }
            justSoldOut:   { type: boolean }
            chefEarnings:  { type: ["number", "null"] }
    OrderDetail:
      type: object
      properties:
        id:             { type: string }
        price:          { type: number }
        effectivePrice: { type: number }
        isMemberOrder:  { type: boolean }
        status:         { type: string }
        orderedAt:      { type: string }
        paymentMethod:  { type: string, enum: [DIGITAL, CASH] }
        escrowStatus:   { type: ["string", "null"], enum: [HELD, RELEASED, CASH, CASH_RECONCILED, CANCELLED, null] }
        pickupToken:    { type: ["string", "null"] }
        dropId:         { type: string }
        dropTitle:      { type: string }
        mealSlot:       { type: string }
        expiresAt:      { type: string }
        inventory:      { type: integer }
        currentOrders:  { type: integer }
        dropStatus:     { type: string }
        pickupLocation: { type: ["string", "null"] }
        chefName:       { type: string }
        userId:         { type: string }
    CancelResult:
      type: object
      required: [id, status]
      properties:
        id:        { type: string }
        status:    { type: string }
        dropId:    { type: ["string", "null"] }
        validUntil: { type: ["string", "null"] }
...